🔧 Fix: Convert NumPy types to Python float for PostgreSQL
✅ Fixed 'schema np does not exist' error
✅ Convert numpy.float64 to Python float before database insert
✅ Ensures PostgreSQL compatibility on Railway

- All numeric values properly converted to native Python types
- Prevents PostgreSQL from interpreting numpy types as schema names

+++ analysis_engine.py
@@ -267,10 +267,10 @@ class AnalysisEngine:
             new_state = {
                 'token_address': token_address,
                 'timeframe': timeframe_str,
-                'high_point': high_point,
-                'low_point': low_point,
-                'target1_price': high_point + (price_range * 0.272),
-                'target2_price': high_point + (price_range * 0.618),
+                'high_point': float(high_point),
+                'low_point': float(low_point),
+                'target1_price': float(high_point + (price_range * 0.272)),
+                'target2_price': float(high_point + (price_range * 0.618)),
                 'status': 'ACTIVE'
             }
             db_manager.upsert_fibo_state(new_state)
